chore: Update total column data type in Invoices table to NUMERIC(10, 2)

+++ api/Migrations/005_CreateInvoicesTable.sql
@@ -4,6 +4,6 @@ CREATE TABLE IF NOT EXISTS noir.Invoices (
     id UUID PRIMARY KEY DEFAULT uuid_generate_v4(),
     account_id UUID NOT NULL REFERENCES noir.Accounts(id),
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    total INT NOT NULL,
+    total NUMERIC(10, 2) NOT NULL,
     status INT NOT NULL,
 );
